% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addLayer.R
\name{addLayer}
\alias{addLayer}
\title{add a layer from a shapefile to CCH portal}
\usage{
addLayer(filename, token)
}
\arguments{
\item{filename}{a shapefile to post}

\item{token}{token received from the authenticateUser function}
}
\value{
layer url for the new layer
}
\description{
returns url of the created layer
}
\details{
This layer is used to track WFS, WMS, and CSW pointers
}

