% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlias.R
\name{aliasExists}
\alias{aliasExists}
\title{alias exists}
\usage{
aliasExists(aliasID)
}
\arguments{
\item{aliasID}{a string for a valid alias identifier}
}
\value{
a boolean
}
\description{
returns whether or not an alias with the specified ID exists
}
\examples{
\dontrun{
setBaseURL('dev')
aliasExists('test')
}
}
