% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authenticateUser.R
\name{authenticateUser}
\alias{authenticateUser}
\title{authenticate a user to CCH portal}
\usage{
authenticateUser(username, password)
}
\arguments{
\item{username}{an active directory username (e.g., \code{'bbadger'})}

\item{password}{a character password for the active directory username.
Will be prompted for password if missing and in interactive mode}
}
\value{
a character token, or the status code if not 200
}
\description{
returns a token that can be used in POST or PUT methods
}
\details{
if \code{authenticateUser} is called with a \code{username} argument,
\code{username} is stored in the R session environment, and future calls to
\code{authenticateUser} within the same R session can use the password argument
only (e.g., \code{authenticateUser(password = '12345')})
}

