% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkItemExists.R
\name{checkItemExists}
\alias{checkItemExists}
\title{Checks to see if an item exists}
\usage{
checkItemExists(itemID)
}
\arguments{
\item{itemID}{id of item to check}

\item{...}{extra parameters to pass to checkAuth}
}
\value{
returns TRUE if the item exists, and false if it does not
}
\description{
returns a boolean success
}
\examples{
\dontrun{
setBaseURL('dev')
checkItemExists('CCGftiy')
}
}
