% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAndValidateAlias.R
\name{cleanAndValidateAlias}
\alias{cleanAndValidateAlias}
\title{clean and validate alias}
\usage{
cleanAndValidateAlias(alias)
}
\arguments{
\item{alias}{the alias to check}
}
\value{
the cleaned alias or NULL if the provided alias is invalid
}
\description{
checks wheter a provided alias is valid and cleans what can be cleaned
}
