% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStorm.R
\name{createStorm}
\alias{createStorm}
\title{Create modeled storm}
\usage{
createStorm(templateId = NULL, filename, aliasName = NULL, ...)
}
\arguments{
\item{templateId}{id of the template to replace}

\item{filename}{a shapefile to post}

\item{aliasName}{(optional) the name of the alias to associate with this storm}

\item{...}{extra params to pass to checkAuth}
}
\value{
character vector representing ID of new template
}
\description{
Create items associated with model run of a storm
}
\details{
This will create all children associated with a storm
and aggregate them under the template.
}
\examples{
\dontrun{
createStorm("CAQw7M1", system.file("extdata", "Sandy_CIDA.zip", "hurricane-sandy",
   package="hazardItems"))
}
}
