% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDownload.R
\name{deleteDownload}
\alias{deleteDownload}
\title{delete the cache for an item download}
\usage{
deleteDownload(itemId, ...)
}
\arguments{
\item{itemId}{id of download to blow away}

\item{...}{extra params to pass to checkAuth}
}
\value{
TRUE if download is deleted
}
\description{
deletes cached download, can then be re-seeded
}
\examples{
\dontrun{
 authenticateUser("bbadger")
 deleteDownload("CCNKrhr")
}
}

