% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTileCache.R
\name{deleteTileCache}
\alias{deleteTileCache}
\title{delete tile cache for application}
\usage{
deleteTileCache(...)
}
\arguments{
\item{...}{extra params to pass to checkAuth}
}
\value{
TRUE if delete worked
}
\description{
destroys existing tile cache
}
\examples{
\dontrun{
 setBaseURL('dev')
 deleteTileCache(username='bbadger')
}
}

