% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlias.R
\name{getAlias}
\alias{getAlias}
\title{get alias}
\usage{
getAlias(aliasID)
}
\arguments{
\item{aliasID}{a string for a valid alias identifier}
}
\value{
a list
}
\description{
returns a list of alias contents
}
\examples{
\dontrun{
setBaseURL('dev')
getAlias('test')
}
}
