% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getItem.R
\name{getItem}
\alias{getItem}
\title{get item or subset of fields}
\usage{
getItem(itemID, field = NULL, subtree = FALSE)
}
\arguments{
\item{itemID}{a string for a valid item identifier}

\item{field}{a subfield to return from the item. Default is NULL}

\item{subtree}{whether or not to return the item subtree}
}
\value{
a list or a character vector (depending on \code{field} argument)
}
\description{
returns a list of item contents
}
\examples{
\dontrun{
setBaseURL('dev')
getItem('CAQw7M1')
getItem('CAQw7M1', field = 'children')
}
}

