% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\title{get layer info}
\usage{
getLayer(layerId)
}
\arguments{
\item{layerId}{a string for a valid layer identifier}
}
\value{
a list representing layer information
}
\description{
returns a list of layer contents
}
\examples{
\dontrun{
setBaseURL('dev')
getLayer('DAozfKaK')
}
}

