% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTemplateItem.R
\name{getTemplateItem}
\alias{getTemplateItem}
\title{get template item or subset of fields}
\usage{
getTemplateItem(itemID, field = NULL)
}
\arguments{
\item{itemID}{a string for a valid item identifier}

\item{field}{a subfield to return from the item. Default is NULL}
}
\value{
a list or a character vector (depending on \code{field} argument)
}
\description{
returns a list of item contents
}
\examples{
\dontrun{
setBaseURL('dev')
getTemplateItem('CAQw7M1')
getTemplateItem('CAQw7M1', field = 'children')
}
}

