% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabXML.R
\name{grabXML}
\alias{grabXML}
\title{create xml document from endpoint}
\usage{
grabXML(serviceEndpoint)
}
\arguments{
\item{serviceEndpoint}{}
}
\value{
XML document
}
\description{
This function takes an endpoint, determines if its local or web-based and returns an xml doc

Title
}
