% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.service.R
\name{historical.service}
\alias{historical.service}
\title{create summary json for an individual historical item}
\usage{
historical.service(serviceEndpoint, attribute)
}
\arguments{
\item{serviceEndpoint}{valid xml, either local or a url}

\item{attribute}{an attribute to use form the shapefile corresponding to 
\code{serviceEndpoint}}
}
\value{
Serialized JSON for summary
}
\description{
This service parses the metadata record and creates a summary specific to the 
HISTORICAL theme, with subtype being defined by the attribute name.
}
\examples{
serviceEndpoint  <-	system.file('extdata',
"NewJerseyS_shorelines.shp.xml", package = 'hazardItems')
attribute	<-	'date'
summary	<-	historical.service(serviceEndpoint,attribute)
print(summary)
sink('output.json')
cat(summary)
sink()
}

