% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Load CCH config}
\usage{
loadConfig(filename)
}
\description{
Loads the config file into options which are
used elsewhere in the application
}
\examples{
loadConfig()
}
\author{
Jordan Walker
}

