% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postItem.R
\name{postItem}
\alias{postItem}
\title{POST item}
\usage{
postItem(item, ...)
}
\arguments{
\item{item}{list representing item to POST}

\item{...}{extra params to pass to checkAuth}
}
\value{
character vector representing id of new item
}
\description{
adds an item to application
}
\examples{
\dontrun{
setBaseURL('dev')
postItem(list(
 itemType="aggregation",
 name="test",
 type="storms"
))
}
}

