% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshItemThumbnail.R
\name{refreshItemThumbnail}
\alias{refreshItemThumbnail}
\title{Refreshes an item's thumbnail PNG}
\usage{
refreshItemThumbnail(itemID, ...)
}
\arguments{
\item{itemID}{id of item to refresh thumbnail for}

\item{...}{extra parameters to pass to checkAuth}
}
\value{
returns TRUE if the thumbnail is put successfully, and false if it fails
}
\description{
returns a boolean success
}
\examples{
\dontrun{
setBaseURL('dev')
refreshItemThumbnail('CCGftiy')
}
}
