% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAlias.R
\name{saveAlias}
\alias{saveAlias}
\title{save alias}
\usage{
saveAlias(alias, isNew, ...)
}
\arguments{
\item{alias}{list representing alias to POST or PUT}

\item{isNew}{whether or not the alias is new}

\item{...}{extra params to pass to checkAuth}
}
\value{
saved alias ID
}
\description{
adds or updates an alias in the application
}
\examples{
\dontrun{
setBaseURL('dev')
saveAlias(
 alias=list("id"="test", "item_id"="test-item"),
 isNew=TURE
)
}
}
