% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveItem.R
\name{saveItem}
\alias{saveItem}
\title{save item}
\usage{
saveItem(item, ...)
}
\arguments{
\item{item}{list representing item to POST or PUT}

\item{...}{extra params to pass to checkAuth}
}
\value{
character vector representing id of new item
}
\description{
adds or updates an item in the application
}
\examples{
\dontrun{
setBaseURL('dev')
postItem(list(
 itemType="aggregation",
 name="test",
 type="storms"
))
}
}

