% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTrack.R
\name{saveTrack}
\alias{saveTrack}
\title{save NHC items}
\usage{
saveTrack(...)
}
\arguments{
\item{...}{extra params to pass to checkAuth}

\item{item}{list representing items to POST or PUT}
}
\value{
TRUE if it worked FALSE otherwise
}
\description{
adds or updates the NHC items in the application
}
\examples{
\dontrun{
setBaseURL('dev')
saveTrack()
}

}

