% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setBaseURL.R
\name{setBaseURL}
\alias{setBaseURL}
\title{Set CCHP endpoint}
\usage{
setBaseURL(endpoint = "prod")
}
\arguments{
\item{endpoint}{Indicate which CCHP endpoint
you want to use options: \code{c('prod','dev','qa')}}
}
\description{
Sets the internal URLS used to either the production, dev or QA server.
URLS are stored internally to the package
}
\examples{
\dontrun{
set_endpoint('prod')
set_endpoint('dev')
}
}
\author{
Luke Winslow, Jordan S Read
}

