% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{instantiate a template with new items based on new data}
\usage{
template(templateId, items, token)
}
\arguments{
\item{templateId}{id of the template to instantiate}

\item{items}{list of items to create (attr or item id and layerId)}

\item{token}{auth token to pass in for templating}
}
\value{
TRUE if it worked FALSE otherwise
}
\description{
creates several items based on data passed to template
}
\details{
This can be used to either create new items from a template
or replace existing items under a template
}
\examples{
# replace existing children items
 items <- list(children=list(
   list(
     id="DvwQhQEH",
     layerId="DvwcwCrK"
   ),
   list(
     id="DvwQjFYJ",
     layerId="DvwcwCrK"
   )
 ))
 # create new children from template
 items <- list(children=list(
   list(
     attr="MEAN",
     layerId="DvwcwCrK"
   ),
   list(
     attr="RUNUP",
     layerId="DvwcwCrK"
   )
 ))
\dontrun{
 template("CAQw7M1", items, authenticateUser(username))
}
}

