% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vulnerability.service.R
\name{vulnerability.service}
\alias{vulnerability.service}
\title{create summary json for an individual vulnerability item}
\usage{
vulnerability.service(serviceEndpoint, attribute)
}
\arguments{
\item{serviceEndpoint}{valid xml, either local or a url}

\item{attribute}{an attribute to use form the shapefile corresponding to
\code{serviceEndpoint}}
}
\value{
Serialized JSON for summary
}
\description{
This service parses the metadata record and creates a summary specific to the
VULNERABILITY theme, with subtype being defined by the attribute name.
}
\examples{
serviceEndpoint  <- paste0('http://cida.usgs.gov/coastalhazards/csw/?',
'service=CSW&request=GetRecordById&version=2.0.2&typeNames=fgdc:metadata',
'&id=urn:uuid:12d6c1ee-a9e6-11e3-99f6-0050569524e0&',
'outputSchema=http://www.opengis.net/cat/csw/csdgm&elementSetName=full')
attribute <- 'DHIGH'
summary <- vulnerability.service(serviceEndpoint,attribute)
print(summary)
}

