% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addLayer.R
\name{addLayer}
\alias{addLayer}
\title{add a layer from a shapefile to CCH portal}
\usage{
addLayer(filename, ...)
}
\arguments{
\item{filename}{a shapefile to post}

\item{...}{additional arguments passed to \code{\link{authenticateUser}}}
}
\value{
layer id for the new layer that was created
}
\description{
returns url of the created layer
}
\details{
This layer is used to track WFS, WMS, and CSW pointers
}
\examples{
\dontrun{
authenticateUser('bbadger')
addLayer(system.file("extdata", "Sandy_CIDA.zip",
   package="hazardItems"))
}
}

