% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadItem.R
\name{downloadItem}
\alias{downloadItem}
\title{download item}
\usage{
downloadItem(itemId, ...)
}
\arguments{
\item{itemId}{id to download}

\item{...}{extra params to pass to checkAuth}
}
\value{
TRUE if download worked
}
\description{
downloads item from cache, or builds cache and gets download
}
\examples{
\dontrun{
 downloadItem("CCNKrhr")
}
}

