% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabXML.R
\name{grabXML}
\alias{grabXML}
\title{create xml document from endpoint}
\usage{
grabXML(serviceEndpoint)
}
\arguments{
\item{serviceEndpoint}{}
}
\value{
XML document
}
\description{
This function takes an endpoint, determines if its local or web-based and returns an xml doc

Title
}
\examples{
serviceEndpoint  <- paste0('https://cida.usgs.gov/coastalhazards/csw/?',
'service=CSW&request=GetRecordById&version=2.0.2&typeNames=fgdc:metadata',
'&id=urn:uuid:12d6c1ee-a9e6-11e3-99f6-0050569524e0&',
'outputSchema=http://www.opengis.net/cat/csw/csdgm&elementSetName=full')
doc <- grabXML(serviceEndpoint)
}

