% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBaseURL.R
\name{setBaseURL}
\alias{setBaseURL}
\title{Set CCHP endpoint}
\usage{
setBaseURL(endpoint = "prod")
}
\arguments{
\item{endpoint}{Indicate which CCHP endpoint 
you want to use options: \code{c('prod','qa','dev')}.
Fully qualified urls are also allowed
(i.e. \code{https://localhost:8080/coastal-hazards-portal/})}
}
\description{
Sets the internal URLS used to either the production, QA, or dev server. 
URLS are stored internally to the package
}
\examples{
\dontrun{
setBaseURL('prod')
setBaseURL('qa')
setBaseURL('dev')
}
}
\author{
Luke Winslow, Jordan S Read
}

