% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateStorm.R
\name{updateStorm}
\alias{updateStorm}
\title{Update storm}
\usage{
updateStorm(templateId, filename, ...)
}
\arguments{
\item{templateId}{id of the template to instantiate}

\item{filename}{a shapefile to post}

\item{...}{extra params to pass to checkAuth}
}
\value{
TRUE if it worked FALSE otherwise
}
\description{
Adds all attributes to an existing template
}
\details{
This will create all children associated with a storm
and aggregate them under the template.
}
\examples{
\dontrun{
updateStorm("CAQw7M1", system.file("extdata", "Sandy_CIDA.zip", 
   package="hazardItems"))
}
}
